UIEUS-526-move-delete-udp-button adapt tests
